requirements.txt
python-dotenv==1.0.0
openai==0.28.1
rich==13.5.2
typer==0.9.0
